pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clawagents"
version = "5.21.0"
description = "Lean, full-stack agentic AI framework with PTRL (Prompt-Time Reinforcement Learning)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "ClawAgents Contributors" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
gemini = ["google-genai>=0.1.0"]
anthropic = ["anthropic>=0.40.0"]
accurate-tokens = ["tiktoken>=0.5.0"]
all = ["google-genai>=0.1.0", "anthropic>=0.40.0", "tiktoken>=0.5.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "tiktoken>=0.5.0",
    "google-genai>=0.1.0",
    "anthropic>=0.40.0",
]

[project.urls]
Homepage = "https://github.com/x1jiang/clawagents_py"
Repository = "https://github.com/x1jiang/clawagents_py"
Issues = "https://github.com/x1jiang/clawagents_py/issues"

[project.scripts]
clawagents = "clawagents.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["benchmark: performance benchmark tests (deselect with -m 'not benchmark')"]
